+++ hutool-http/src/main/java/cn/hutool/http/HttpResponse.java
@@ -1,5 +1,6 @@
 package cn.hutool.http;
 
+import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.convert.Convert;
 import cn.hutool.core.io.FastByteArrayOutputStream;
 import cn.hutool.core.io.FileUtil;
@@ -467,24 +468,60 @@ public class HttpResponse extends HttpBase<HttpResponse> implements Closeable {
 
 	/**
 	 * 从Content-Disposition头中获取文件名
-	 * @param paramName 文件参数名
 	 *
+	 * @return 文件名，empty表示无
+	 */
+	public String getFileNameFromDisposition() {
+		return getFileNameFromDisposition(null);
+	}
+
+	/**
+	 * 从Content-Disposition头中获取文件名，以参数名为`filename`为例，规则为：
+	 * <ul>
+	 *     <li>首先按照RFC5987规范检查`filename*`参数对应的值，即：`filename*="example.txt"`，则获取`example.txt`</li>
+	 *     <li>如果找不到`filename*`参数，则检查`filename`参数对应的值，即：`filename="example.txt"`，则获取`example.txt`</li>
+	 * </ul>
+	 * 按照规范，`Content-Disposition`可能返回多个，此处遍历所有返回头，并且`filename*`始终优先获取，即使`filename`存在并更靠前。<br>
+	 * 参考：https://developer.mozilla.org/zh-CN/docs/Web/HTTP/Headers/Content-Disposition
+	 *
+	 * @param paramName 文件参数名，如果为{@code null}则使用默认的`filename`
 	 * @return 文件名，empty表示无
 	 */
 	public String getFileNameFromDisposition(String paramName) {
 		paramName = ObjUtil.defaultIfNull(paramName, "filename");
+		final List<String> dispositions = headerList(Header.CONTENT_DISPOSITION.name());
 		String fileName = null;
-		final String disposition = header(Header.CONTENT_DISPOSITION);
-		if (StrUtil.isNotBlank(disposition)) {
-			fileName = ReUtil.get(paramName+"=\"(.*?)\"", disposition, 1);
-			if (StrUtil.isBlank(fileName)) {
-				fileName = StrUtil.subAfter(disposition, paramName + "=", true);
+		if (CollUtil.isNotEmpty(dispositions)) {
+
+			// filename* 采用了 RFC 5987 中规定的编码方式，优先读取
+			fileName = getFileNameFromDispositions(dispositions, StrUtil.addSuffixIfNot(paramName, "*"));
+			if ((!StrUtil.endWith(fileName, "*")) && StrUtil.isBlank(fileName)) {
+				fileName = getFileNameFromDispositions(dispositions, paramName);
 			}
 		}
+
 		return fileName;
 	}
 
 	// ---------------------------------------------------------------- Private method start
+
+	/**
+	 * 从Content-Disposition头中获取文件名
+	 *
+	 * @param dispositions Content-Disposition头列表
+	 * @param paramName    文件参数名
+	 * @return 文件名，empty表示无
+	 */
+	private static String getFileNameFromDispositions(final List<String> dispositions, String paramName) {
+		String fileName = null;
+		for (String disposition : dispositions) {
+			fileName = ReUtil.getGroup1(paramName + "=\"(.*?)\"", disposition);
+			if (StrUtil.isNotBlank(fileName)) {
+				break;
+			}
+		}
+		return fileName;
+	}
 
 	/**
 	 * 初始化Http响应，并在报错时关闭连接。<br>
@@ -632,7 +669,7 @@ public class HttpResponse extends HttpBase<HttpResponse> implements Closeable {
 		} catch (IORuntimeException e) {
 			//noinspection StatementWithEmptyBody
 			if (isIgnoreEOFError
-					&& (e.getCause() instanceof EOFException || StrUtil.containsIgnoreCase(e.getMessage(), "Premature EOF"))) {
+				&& (e.getCause() instanceof EOFException || StrUtil.containsIgnoreCase(e.getMessage(), "Premature EOF"))) {
 				// 忽略读取HTTP流中的EOF错误
 			} else {
 				throw e;